Merged PR 4742: Support per-binding protocol
We need to support proxying connections via YARP where the backend also speaks HTTPS.

This is required to support the webpack development server that the frontend uses for a few reasons:

- The webpack server does not follow `X-Forwarded-Proto` directives so if you disable https in the webpack server, the links in the pages that are served will be plain http and cause mixed content problems or worse
- Localhost being considered a secure origin only works for localhost and not `workleap.local` so using secure-only apis requires the server to be running with https enabled.

Note that proxying the frontend also requires changes to the webpack `devServer` configuration so that links are rewritten to the leap proxy's address rather than the backend address of the `devServer`. The `public_url` must point to the leap proxy url.