Merged PR 4932: IDP-939 Don't publish service discovery env variables to dependencies
It turns out that this was actually already implemented in a way but due to a typo, this service discovery information was leaking into the mongodb container. I plan to clean up how this information is passed around in the future but this was a great insight into the internals of the Leap code base for me. I have simply fixed the typo and confirmed that mondodb no longer gets re-created when adding or removing docker services in the leap stack.