Merged PR 5466: IDP-1089 - Changed leap to run Local env and have extensions package
Changed Leap to run on environment Local and not Development since Development is already in use within the platform. Realized that using Development could disrupt current development flow and would force everyone who want to use those services to opt-in to Leap, which is not what we want.

Also added a new extensions package to offer an extension method that will allow services to add the generated appsettings.json to their configuration. This will make it easier for services running in an IDE to communicate with services running in Leap. This is a usecase we forsee happening.

I tested the extension package with a locally built version