Merged PR 5502: IDP-1129 - Added tags as a publish trigger
A past PR changed things so we would use GitVersion and tags for Leap, but the tag trigger was forgotten from the publish pipeline so currently tagging doesn't work as a way to publish a new version. That's why it was added to the publish pipeline.

Also changed the `assembly-informational-format` in `GitVersion.yml` from `NugetVersion` to `NuGetVersion` since that's how it's written in the doc https://gitversion.net/docs/reference/variables