Merged PR 5631: IDP-1160 Add FusionAuth dependency to Leap
Add FusionAuth dependency to Leap. This was a bit tricky because Leap itself depends on PostgreSQL to store data, so a mechanism for dependencies to declare children was added. This mechanism isn't super robust for reasons explained in code comments but I judge that it is good enough for now.

The fusionauth instance is kickstarted with the same settings as the docker-compose for the fusionauth provisioning project so it should be very familiar to end-users.