Merged PR 5793: [IDP-1212] Ensure hosted web apps are stopped before disposing resources
We used to simply dispose the kestrel-based web apps that we start in leap: aspire, yarp, az cli proxy for docker.

The issue with Aspire was that disposing the DistributedApplication instance (and the underlying host) would not stop incoming traffic from getting in from the blazor dashboard. This traffic would trigger code paths that want to use already disposed resources, such as the LoggerFactory as seen in the Jira screenshot.

The fix is to stop the web app first, then dispose it. The exception is gone.

I didn't do the change for az cli proxy for docker because the app is hosted on aspire, if ever there was an exception, nobody could see it as the dashboard would no longer be available - and that's the only place where the az cli proxy's logs can be found.