Merged PR 5890: [IDP-1079] Download matching .NET Aspire workload at runtime
## The .NET Aspire workload is no longer required -> users won't have to run "dotnet workload install/update aspire".

We download the Aspire SDK NuGet packages that are part of the workload at runtime (orchestrator and dashboard) using the NuGet client libraries.
- The NuGet packages are stored in `~/.leap/generated/nuget-packages/`
- The NuGet global package cache is used (skip download when possible)
- If Aspire NuGet packages already there > we do nothing

The Aspire SDK NuGet packages' version are now matching the NuGet hosting package. The version is embedded in the assembly at build, and we retrieve it at runtime. It's a more reliable technique than using reflection to extract the assembly version on an Aspire type.

Try the generated package by doing this first:
- Run `dotnet workload uninstall aspire` to make sure you don't have the workload
- Delete your `~/.leap/` dir to start fresh
- Eventually delete Aspire-related packages in your NuGet global package cache

Also now using the Aspire.Hosting.AppHost package instead of Aspire.Hosting which is available since preview 6 and has less dependencies