Merged PR 6579: [IDP-1884] Support Docker bind mounts in Docker runners
Example:

```yaml
services:
  myapp:
    ingress:
      host: myapp.workleap.localhost
    runners:
      - type: docker
        image: mcr.microsoft.com/dotnet/samples:aspnetapp
        containerPort: 8080
        volumes:
          - "../../../shared_stuff:/data/" # here, shared_stuff is a directory, but it also works for files
```

As for any path host path in `leap.yaml`, .`./shared_stuff` (the mapping source) will be normalized relatively to the location of the containing `leap.yaml` file. The syntax comes from Docker Compose.

Proof that it works:

![image (3).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/6579/attachments/image%20%283%29.png)

If the file or directory does not exists, a warning will be shown:

![image (2).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/6579/attachments/image%20%282%29.png)

Other errors has been handled as well:
- Empty source or destination mapping
- Destination mapping is not rooted (absolute path)

In order to implement this, I had to split the huge method into smaller ones (see `PopulateServicesFromYamlPipelineStep`).