Merged PR 7100: [IDP-2177] Define ASP.NET Core default certificate when protocol is https
## Context

When a developer defines the protocol as `https`, it means the backend is expected to use HTTPS with a certificate. While developers could manually set the certificate, we can automatically do it for .NET apps (dotnet run and docker images). It is less work for them overall, and they can have end-to-end HTTPS from the reverse proxy to the app running on localhost.

## QA

I enabled the protocol `https` for several .NET web apps using docker, executable, and dotnet runners. Browsing to the `https://localhost:*` app have shown that we use our mkcert cert by default without additional configuration.