Merged PR 7597: [IDP-2458] Dotnet and executable runners wait for AzureCliProxy on fixed port
We've noticed that services may start before AzureCliProxy has started. That causes an issue where the user would need to restart the service manually. Given that Aspire 9.0 offers a WaitFor implementation, we can now configure our services start only after AzureCliProxy is in a Running state.

Additionally, we are updating AzureCliProxy to start on a fixed port: 6501