Merged PR 7628: [IDP-2228] Support user runner preferences for service
Leap users can define more than one runner preference in their leap.yaml config files. However, we previously only picked the first one in the list as the default runner. If they wanted to run Leap with a different runner, they would need to update their Leap files's runner order. This causes friction in the user experience.

As such, we are introducing a new Leap command: `preferences` with 3 subcommands:

```
leap preferences set --service <service name> --runner <runner name>

leap preferences remove --service <service name>

leap preferences list
```
![image.png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7628/attachments/image.png)
### 1. preferences set
Required options:
`--service, -s`: Service name for which to set a preference.
`--runner, -r`: Runner name for the preference.

This command allows users to set preferences for a service. Even if a previous preference is declared, this command is overwrite the previous preference.
![image (2).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7628/attachments/image%20%282%29.png)

### 2. preferences remove
Required options:
`--service, -s`: Service name for which to set a preference.

This command allows users to remove preferences for a service.
![image (3).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7628/attachments/image%20%283%29.png)

### 3. preferences list
This command lists preferences of all services defined by the user.
![image (4).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7628/attachments/image%20%284%29.png)

Once the preferences are set, the `PopulateServicesFromYamlPipelineStep` will first check if a preference was declared, then process that service. If no preferences are declared, we will default to the first runner.

The preferences are saved to `%PATH%/.leap/leappreferences.json`

## QA scenarios

1. Test set/remove/list commands with valid & invalid inputs. Verify that they are written to file as expected and that validation catches bad inputs. - Done

2. Leap.yaml with 2 different runners with preference defined for 2nd runner. Validate that 2nd runner is picked.
![image (7).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7628/attachments/image%20%287%29.png)

2. Leap.yaml with 2 different runners with no preference. Validate that 1st runner is picked with log.
![image (8).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7628/attachments/image%20%288%29.png)

3. Leap.yaml with 2 different runners wi...