Merged PR 7661: [IDP-2494] Print dashboard URL as soon as it is up
With the introduction of WaitFor in .NET Aspire 9, the distributed application waits for resource dependencies to be available before returning when we do `app.StartAsync()`.

Because we display the dashboard URL only when the app is started, it might take users a while to become aware that the dashboard is available.

This PR aims to show the dashboard URL as soon as it is ready. Also includes a few other changes described in PR comments.

## QA

I've put a 20 seconds delay in `StartAzureCliDockerProxyPipelineStep.HostAzureCliDockerProxyInAspireLifecycleHook.TrackAzureCliProxyResource` and prior to this change, the Leap CLI would not show the dashboard URL until the delay was completed. The dashboard was already available though.

Now, with these changes and the delay, we can see the Azure CLI proxy resource as "waiting" and the dashboard URL is printed as soon as possible:

![image (4).png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7661/attachments/image%20%284%29.png)