Merged PR 7659: [IDP-2472] Dotnet and Executable resources WaitFor all dependencies before start
[IDP-2472] Dotnet and Executable resources WaitFor all dependencies before start

Previously, services were only waiting for the AzureCli resource to start. However, it would make sense to have the service wait for all dependencies to be started instead. As such, this PR changes the Dotnet and Executable resources to wait for all dependencies to be in a `Running` state before starting.

QA:
Confirmed that we are waiting for the dependencies to start.
![image.png](https://dev.azure.com/gsoft/46c89666-c1f6-4351-a213-a0a678ed8fe7/_apis/git/repositories/92e69d95-09ff-48b8-8197-a9831c6bf49b/pullRequests/7659/attachments/image.png)